Added websocket.js file and modified studio.tsx
