+++ backend/index.js
@@ -1,7 +1,9 @@
 
 import app from './app.js'
 import { connectToDatabase } from './db/connections.js'
+import WebSocket from './utils/websocket.js';
 
+WebSocket();
 connectToDatabase()
     .then(() => {
         app.listen(8000, () => console.log("server started and connected to db at port 8000"));
