+++ backend/models/schema.js
@@ -10,8 +10,8 @@ const userSchema = new Schema({
   name: { type: String, required: true },
   email: { type: String, required: true },
   password: { type: String, required: true },
-  destinationSchema:[destinationSchema],
-  broadcastSchema:[broadcastSchema],
+  //destinationSchema:[destinationSchema],
+  //broadcastSchema:[broadcastSchema],
 },{
     timestamp: true,
 });
